feat: Refactor helper function to generate started and finished events
