+++ pkg/lib/v0_2_0/error.go
@@ -1,0 +1,12 @@
+package v0_2_0
+
+type Error struct {
+	StatusType StatusType
+	ResultType ResultType
+	Message    string
+	Err        error
+}
+
+func (e Error) Error() string {
+	return e.Message
+}
