+++ pkg/sdk/keptn_test.go
@@ -3,9 +3,10 @@ package sdk
 import (
 	"context"
 	"fmt"
-	"github.com/keptn/go-utils/pkg/sdk/internal/config"
 	"math"
 	"testing"
+
+	"github.com/keptn/go-utils/pkg/sdk/internal/config"
 
 	"github.com/google/uuid"
 	"github.com/keptn/go-utils/pkg/api/models"
@@ -42,7 +43,7 @@ func Test_NewKeptn(t *testing.T) {
 
 func Test_ReceivingInvalidEvent(t *testing.T) {
 	taskHandler := &TaskHandlerMock{}
-	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error) { return FakeTaskData{}, nil }
+	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error) { return FakeTaskData{}, nil }
 	fakeKeptn := NewFakeKeptn("fake")
 	fakeKeptn.AddTaskHandler("sh.keptn.event.faketask.triggered", taskHandler)
 	fakeKeptn.NewEvent(models.KeptnContextExtendedCE{
@@ -103,7 +104,7 @@ func Test_SendEvents(t *testing.T) {
 
 func Test_ReceivingEventWithMissingType(t *testing.T) {
 	taskHandler := &TaskHandlerMock{}
-	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error) { return FakeTaskData{}, nil }
+	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error) { return FakeTaskData{}, nil }
 	fakeKeptn := NewFakeKeptn("fake")
 	fakeKeptn.AddTaskHandler("sh.keptn.event.faketask.triggered", taskHandler)
 	fakeKeptn.NewEvent(models.KeptnContextExtendedCE{
@@ -118,7 +119,7 @@ func Test_ReceivingEventWithMissingType(t *testing.T) {
 
 func Test_CannotGetEventSenderFromContext(t *testing.T) {
 	taskHandler := &TaskHandlerMock{}
-	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error) { return FakeTaskData{}, nil }
+	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error) { return FakeTaskData{}, nil }
 	fakeKeptn := NewFakeKeptn("fake")
 	fakeKeptn.AddTaskHandler("sh.keptn.event.faketask.triggered", taskHandler)
 	fakeKeptn.Keptn.OnEvent(context.TODO(), models.KeptnContextExtendedCE{
@@ -133,7 +134,7 @@ func Test_CannotGetEventSenderFromContext(t *testing.T) {
 
 func Test_WhenReceivingAnEvent_StartedEventAndFinishedEventsAreSent(t *testing.T) {
 	taskHandler := &TaskHandlerMock{}
-	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error) { return FakeTaskData{}, nil }
+	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error) { return FakeTaskData{}, nil }
 	fakeKeptn := NewFakeKeptn("fake")
 	fakeKeptn.AddTaskHandler("sh.keptn.event.faketask.triggered", taskHandler)
 	fakeKeptn.NewEvent(models.KeptnContextExtendedCE{
@@ -210,8 +211,8 @@ func Test_WhenReceivingAnEvent_AndAutomaticEventResponseIsDisabledOnTaskHandler_
 
 func Test_WhenReceivingAnEvent_TaskHandlerFails(t *testing.T) {
 	taskHandler := &TaskHandlerMock{}
-	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error) {
-		return nil, &Error{
+	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error) {
+		return nil, &v0_2_0.Error{
 			StatusType: v0_2_0.StatusErrored,
 			ResultType: v0_2_0.ResultFailed,
 			Message:    "something went wrong",
@@ -237,7 +238,7 @@ func Test_WhenReceivingAnEvent_TaskHandlerFails(t *testing.T) {
 
 func Test_WhenReceivingBadEvent_NoEventIsSent(t *testing.T) {
 	taskHandler := &TaskHandlerMock{}
-	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error) { return FakeTaskData{}, nil }
+	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error) { return FakeTaskData{}, nil }
 	fakeKeptn := NewFakeKeptn("fake")
 	fakeKeptn.AddTaskHandler("sh.keptn.event.faketask.triggered", taskHandler)
 	fakeKeptn.NewEvent(newTestTaskBadTriggeredEvent())
@@ -246,7 +247,7 @@ func Test_WhenReceivingBadEvent_NoEventIsSent(t *testing.T) {
 
 func Test_WhenReceivingAnEvent_AndNoFilterMatches_NoEventIsSent(t *testing.T) {
 	taskHandler := &TaskHandlerMock{}
-	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error) { return FakeTaskData{}, nil }
+	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error) { return FakeTaskData{}, nil }
 	fakeKeptn := NewFakeKeptn("fake")
 	fakeKeptn.AddTaskHandler("sh.keptn.event.faketask.triggered", taskHandler, func(keptnHandle IKeptn, event KeptnEvent) bool { return false })
 	fakeKeptn.NewEvent(models.KeptnContextExtendedCE{
@@ -262,7 +263,7 @@ func Test_WhenReceivingAnEvent_AndNoFilterMatches_NoEventIsSent(t *testing.T) {
 
 func Test_NoFinishedEventDataProvided(t *testing.T) {
 	taskHandler := &TaskHandlerMock{}
-	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error) {
+	taskHandler.ExecuteFunc = func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error) {
 		return nil, nil
 	}
 	fakeKeptn := NewFakeKeptn("fake")
@@ -336,10 +337,10 @@ type FakeTaskData struct {
 }
 type TaskHandlerMock struct {
 	// ExecuteFunc mocks the Execute method.
-	ExecuteFunc func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error)
+	ExecuteFunc func(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error)
 }
 
-func (mock *TaskHandlerMock) Execute(keptnHandle IKeptn, event KeptnEvent) (interface{}, *Error) {
+func (mock *TaskHandlerMock) Execute(keptnHandle IKeptn, event KeptnEvent) (interface{}, *v0_2_0.Error) {
 	if mock.ExecuteFunc == nil {
 		panic("TaskHandlerMock.ExecuteFunc: method is nil but taskHandler.Execute was just called")
 	}
